requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
tavily-python==0.5.0
pydantic==2.10.0
python-dotenv==1.0.1
vercel==0.0.6
